Re-create dogu admin user at subsequent start; #5
Remove and re-create the sonarqubedoguadmin user at a subsequent dogu start,
because it might have been changed when someone with the user login
"sonarqubedoguadmin" logs in. This would change the users password, which
then would not be the same as the on saved in the registry. It also might
lead to the user not having admin permissions any more.
Re-creating the user resolves all these problem. Additionally, a user with
the name "sonarqubedoguadmin" can log in without any problem and will be
assigned the correct permissions.