Create dogu admin and use it as soon as possible at first start
This change is neccessary, because the default "admin" user may be altered
by a login via the cas plugin which leads to the initialization steps at first start
to fail.
If a CES user has logged in in another dogu already and tries to browse
sonar while it is starting for the first time, the cas-plugin may handle the login
process before the steps that need the default admin:admin user configuration have
been performed, which then leads to the dogu failure described above.