#30 Remove permissions from previous admin group if group name changed
This commit introduces the detection of the change of the admin group.
Since the changed admin group name is unknown to SonarQube, the script
creates a new admin group via the usual SonarQube means. Once this has
happened the original admin group will lose all its permissions.

The sequence of all these permission changes are selected to add more
robustness if adding the new admin group fails (f. i. the name contains
weird characters or whatnot) This way there is always a admin group
which contains the necessary permissions.