#30 extract function for granting admin permissions
Previously, when the admin group was changed or removed the admin group
was re-created along with permissions to be set. This lacked vital
permissions, that is: only 1 of 4 permission was effectively set.

This commit extracts the full set of admin permissions which are granted
the admin group. This includes that now all of the mentioned permissions
are granted.