#46 Fixes faulty IFS usage
IFS is a OS-level environment variable that determines which separator is used
for loop iteration input. This commit stores the original value before any
change occurs and restores it before the method is left.

This fix is necessary because a change of method order will defect other
functions (like `wait_for_sonar_to_get_up()`) which rely on a space as
separator.