#46 Change order for quality profile import, add conditional start/stop
This commit fixes a faulty quality profile import. If any new plugins are
installed SonarQube requires a reboot activate the plugins. These plugins
may refer to functionality referenced in the quality profiles.

If the newly installed plugins are not activated and quality profiles are
imported, the quality profiles end up as an empty descriptor on SonarQubes'
page "Quality Profiles".

This commit specifially **avoids** to use the inistial admin user that is
created during a fresh SonarQube installation. Only a "subsequent"-admin has
sufficient privileges to import quality profiles.

Starting and stopping SonarQube is a time-expensive action. This commit also
reduces the downtime by checking of quality profiles are supposed to be
imported. If not, the usual start-up process is accelerated by avoiding the
unnecessary stop/start action.