#3 Add ces-admin group to default project permission template
Previously, ces-admins did not automatically receive permissions to administer
a SonarQube project. They had to add their user or their admin group to the
project (or apply a customized permission template with similar permissions).

This commit adds the ces-admin group to the default permission template with
all existing project-level permissions if (and only if) the default permission
template exists. All newly created projects that match the template's project
pattern (usually any new project) will receive all permissions for the
ces-admin group.

Since there is no direct relationship between permission template and project
this does not change the permissions of existing projects.

This change does also not apply to those projects which do not match the
project pattern. That means, if the default permission template was changed
to a pattern XYZ and a new project ABC is created then this pattern does not
apply to the newly created project and ces-admins are not automatically enabled
to administer the new project.