Decrease image size by 271 MB
This hotfix commit moves a recursive `RUN chown` command into the previous `RUN`
command and thus shaves 271 MB away from the previously 1.1 GB image.

The `RUN chown` command creates a separate layer and modifies every single file
and directory under `/opt/sonar`. Since the data there is abundant the resulting
image layer is unneccessary large.

Because the resource `COPY` command (that followed said `RUN` command) copies
also files to `/opt/sonar` it was amended with a `--chown` option as well. This
will work out overall because the container run-user is the same user that
copied the resources to `/`