#63 Add notification for upgrade to 8.9.x, add shell tests
Testing the value in /proc/sys/vm/max_map_count is a tough task because
some major Linux container magic needs to happen in order to change the value.

Instead of magic, the max_map_count is saved into a local variable and later
overwritten by an exported variabled inside the respective tests.