#103 Restructure docs for better understandability
The previous documentation lacked a understandable structure because it
was written from the get-go. It also missed some crucial parts.

This commit restructures this documentation so developers unfamiliar
with the configuration of Jenkins and SonarQube in terms of the SQ
community branch plugin can better follow.

This code was almost fully written without AI (except the english
translation, sorry about that).

Co-authored-by: Florian Deeken <florian.deeken@cloudogu.com>