#136 Add go-specific Makefile to sonarcarp for Go compilation
It is desirable to have all test reports (dogu, bats, go, etc) in one
directory: the projects target/ dir.
With the compilation of sonarcarp in sonarcarp we have to reference this
directory relatively ../. This conflicts with the dogu Makefile.

This commit adds a second Makefile and relies on the existance of the
required Makefile include files.

Co-authored-by: Simon Klein <simon.klein@cloudogu.com>