#136 Fix buggy REST request handling, remove superfluous debug output
In the previous refactoring, a necessary request modification PRIOR to
forwarding to the request proxy handler was missing. While it remains
unclear that Host and Scheme must be manually set, it is vital for all
proxied requests. This commit sets these for ALL requests (t. i. not
only for browser requests)

Without the manual modifications the forwarder will return with an
'unsupported scheme ""' which ultimately results in understandable HTTP
500 response codes.

Co-authored-by: Simon Klein <simon.klein-extern@cloudogu.com>