#136 Avoid resetting the throttle limiter after unauthenticated resource
The previous code allowed to exploit a loop-hole in which the limiter is
reset after a single response status code returned okay-ish.
This commit filters superficialy the server-version endpoint as example
to show that the limiter will not reset immediately when a sonarscanner
CLI is tried against the Sonar API with a wrong sqa token.

In the long run it is probably a better idea to check if there are any
static routes (like CSS, JS etc) and re-route them right away without
using the auth middleware at all. This should include unauthenticated
endpoints like `$sonarURL/api/server/version`