#136 Implement static resource handling w/o authentication/throttling
Except DoS'ing the application, there is probably no other profound
problem with allowing requests for static resources.

The idea to always allow static resources is grounded in the assumption
of checking such requests against the CAS and/or resetting the
throttling status after a successful response to a static resource like
a Javascript file.

To avoid unnecessary calls towards the CAS service as well as avoiding
to always reset the throttling after one single successful JS fetch,
this commit always allows fetching resources that are marked as static
resource in the configuration.