#136 Add backchannel session logout infra
The session map maps cas usernames to user objects which are filled
during runtime. This map must be used to fill and log of when the back
channel logout event occurs. The event information of the user is posted
by CAS as SAML information which is still needed to be caught and acted
upon.