#136 Handle http header hacker scenario
It is absolutely in the realm of the possible that an attacker may try
to add additional X headers which may transported through the CES nginx
reverse proxy to circumvent authentication and/or throttling.

This detection must occur pretty early and also before any
unauthenticated routes may be hit (which in turn may SonarQube make set
a valid session cookie). Since it takes some effort to add such header
fields, it has a high probability of an attack. Noone else besides us
uses these header fields during authentication.

An interesting alternative would be to randomly generate the header
fields for both the SonarQube properties and the CARP config during the
container start. This way the attack surface would drastically shrink
for hackers even if they analyze our open source code because the value
is random and no longer fixed.