#144 make go-cas library logging configurable
This code slipped through the previous code review.

The code in question serves to configure logging in go-cas
which depends on flags. To activate we artificially work on the flags by
hand. It should not be on by default because user attributes will be
logged which defies data security rules like GDPR.